app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.9
gdown==5.2.0
